Update for StockRepository
