+++ code/FinanceManager.Domain/Repositories/IStockRepository.cs
@@ -5,6 +5,6 @@ namespace FinanceManager.Domain.Repositories
     public interface IStockRepository
     {
         Task<StockPrice> AddStockPrice(string ticker, decimal pricePerUnit, string currency, DateTime date);
-        Task<StockPrice> GetStockPrice(string ticker, DateTime Date);
+        Task<StockPrice?> GetStockPrice(string ticker, DateTime Date);
     }
 }
